fix image build script

+++ source/images/shrink_img.sh
@@ -9,19 +9,25 @@ echo $0
 # To convert a single image - example:
 # convert site-header-img.jpg -resize 1000x1000\> -quality  80% site-header-img.jpg
 
-find $1 -type f -name '*' -exec sh -c '
+find $1 -type f -exec sh -c '
 	filename=$(basename "$0")
-	extension="${filename##*.}"
-	filename="${filename%.*}"
-	dir=$(dirname "$0")
+	size=$(wc -c <"$0")
+	file_is_big=$(($size > 1000000))
 
-	echo "Shrinking...$0 to $filename.jpg"
-
-	convert "$0" -resize 1700x1700\> -quality  70% "$dir/$filename.jpg"
-
-	if [ "$extension" != "jpg" ]
+	if [ $file_is_big != 0 ]
 	then
-		echo "DELETING...$0"
-		rm "$0"
+		extension="${filename##*.}"
+		filename="${filename%.*}"
+		dir=$(dirname "$0")
+
+		echo "Shrinking...$0 to $filename.jpg"
+
+		convert "$0" -resize 1700x1700\> -quality  70% "$dir/$filename.jpg"
+
+		if [ "$extension" != "jpg" ]
+		then
+			echo "DELETING...$0"
+			rm "$0"
+		fi
 	fi
 ' {} ';'
